// x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_response_actions/endpoint/runscript_config/types.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License
 * 2.0; you may not use this file except in compliance with the Elastic License
 * 2.0.
 */

export interface ValidationState {
  isValid: boolean;
  errors?: string[];
}